Combine Header and ScreenHeader, use HeaderButton
