--- packages/mobile/src/components/header-button/index.ts
@@ -1,1 +1,0 @@
-export { HeaderButton } from './HeaderButton'
