--- packages/mobile/src/components/screen-header/index.ts
@@ -1,1 +1,0 @@
-export { ScreenHeader } from './ScreenHeader'
